pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "specs-agent"
version = "0.1.0"
description = "API Testing TUI - Generate test plans from Swagger/OpenAPI specs"
readme = "README.md"
license = "MIT"
requires-python = ">=3.11"
dependencies = [
    "textual>=1.0.0",
    "prance>=23.6.0",
    "openapi-spec-validator>=0.7.0",
    "httpx>=0.27.0",
    "pyyaml>=6.0",
    "jsonschema>=4.20.0",
    "jinja2>=3.1.0",
    "click>=8.1.0",
    "faker>=30.0.0",
    "hdrhistogram>=0.10.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
]
pdf = [
    "weasyprint>=60.0",
]

[project.scripts]
specs-agent = "specs_agent.__main__:main"

[tool.hatch.build.targets.wheel]
packages = ["src/specs_agent"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
